// Sviluppo App Angular/AppF1/src/app/shared/interfaces/session.ts
export interface session {
    circuit_key: number
    circuit_short_name: string
    country_code: string
    country_key: number
    country_name: string
    date_end: string
    date_start: string
    gmt_offset: string
    location: string
    meeting_key: number
    session_key: number
    session_name: string
    session_type: string
    year: number
  };